@@ -43,7 +43,7 @@
     "electron-store": "^8.1.0",
     "electron-updater": "^5.3.0",
     "element-ui": "^2.15.13",
-    "i18next": "^22.4.12",
+    "i18next": "^22.4.13",
     "lodash": "^4.17.21",
     "node-fetch": "^2.6.1",
     "normalize.css": "^8.0.1",
@@ -85,7 +85,6 @@
     "eslint": "^7.32.0",
     "eslint-friendly-formatter": "^4.0.1",
     "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-n": "^15.6.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-vue": "^9.9.0",
@@ -101,9 +100,9 @@
     "vue-loader": "^15.10.1",
     "vue-style-loader": "^4.1.3",
     "vue-template-compiler": "^2.7.14",
-    "webpack": "^5.76.2",
+    "webpack": "^5.77.0",
     "webpack-cli": "^5.0.1",
-    "webpack-dev-server": "^4.12.0",
+    "webpack-dev-server": "^4.13.1",
     "webpack-hot-middleware": "^2.25.3",
     "webpack-merge": "^5.8.0",
     "worker-loader": "^3.0.8"
