@@ -120,16 +120,20 @@ function greeting () {
   const cols = process.stdout.columns
   let text = ''
 
-  if (cols > 85) text = 'lets-build'
-  else if (cols > 60) text = 'lets-|build'
-  else text = false
+  if (cols > 85) {
+    text = 'lets-build'
+  } else if (cols > 60) {
+    text = 'lets-|build'
+  } else {
+    text = false
+  }
 
   if (text && !isCI) {
     say(text, {
-      colors: ['yellow'],
+      colors: ['magentaBright'],
       font: 'simple3d',
       space: false
     })
-  } else console.log(chalk.yellow.bold('\n  lets-build'))
+  } else console.log(chalk.magentaBright.bold('\n  lets-build'))
   console.log()
 }
